Revert "Merge pull request #12 from Jvtrovao/overworld-expansion"
This reverts commit a714a181d27aabe6c1483c465c46b4f247f0b8bb, reversing
changes made to 138e10134308378898ecfdc3fa599cb2f6ef6dbe.

+++ src/field_specials.c
@@ -1558,7 +1558,7 @@ void ResetContextNpcTextColor(void)
 
 u8 ContextNpcGetTextColor(void)
 {
-    u16 gfxId;
+    u8 gfxId;
     if (gSpecialVar_TextColor != NPC_TEXT_COLOR_DEFAULT)
     {
         // A text color has been specified, use that
